+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Autowiring with non-instantiable class
+- Instantiator with non-instantiable class
 
 ## [2.3.0] - 2024-03-26